tuyul-trading-swarm/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tuyul-trading-swarm"
version = "3.0.0"
description = "Production-grade multi-agent AI trading decision system — Tuyul Exception v.3"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "redis[hiredis]>=5.0.0",
    "loguru>=0.7.2",
    "PyJWT>=2.8.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "aiofiles>=23.2.1",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.0",
    "black>=24.4.2",
    "flake8>=7.0.0",
    "mypy>=1.10.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
tuyul-swarm = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "api*", "core*", "schemas*", "infrastructure*", "storage*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
